f_d_m_p/pyproject.toml
[project]
name = "f-d-m-p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "pymongo>=4.16.0",
    "python-jose>=3.5.0",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "user-agents>=2.2.0",
    "uvicorn>=0.44.0",
]
